Remove `@key` from `products.graphql`
# What changed?
This PR proposes that we remove `@key` from `type Product` to avoid distracting a new user with the cognitive weight of understanding entities from day one. 

# Message for reviewers
I'd love to get a sense of how to go about testing that this doesn't break anything. I'd also love to know if there are any additional steps I should be taking in this PR to avoid said breaking changes. 🙏🏽 

Signed-off-by: Ana De Magalhães <88685630+ana-demags@users.noreply.github.com>